Move webpack stuff indo webpack/
