+++ webpack/startWebpackDevServer.js
@@ -3,7 +3,7 @@ import { info, error } from 'common/nordLogger'
 export default function startWebpackDevServer() {
   const webpack = require('webpack') // eslint-disable-line global-require
   const WebpackDevServer = require('webpack-dev-server') // eslint-disable-line global-require
-  const config = require('../../webpack.local.config') // eslint-disable-line global-require
+  const config = require('./webpack.local.config') // eslint-disable-line global-require
 
   new WebpackDevServer(webpack(config), {
     publicPath: config.output.publicPath,
